pyproject.toml
[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[project]
name = "perplexity-webui-scraper"
version = "0.1.2"
description = "Perplexity WebUI Scraper"
readme = "README.md"
authors = [
    { name = "henrique-coder", email = "henriquemoreira10fk@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "httpx",
    "orjson",
    "pydantic>=2.0.0",
    "fastapi",
]

[dependency-groups]
dev = [
    "camoufox",
    "fastapi",
    "python-dotenv",
    "rich",
    "uvicorn",
]
lint = [
    "ruff",
]
